create Storage class for the chatBot to load and save tasks in the file
