+++ src/main/java/Task.java
@@ -1,11 +1,7 @@
 import java.io.*;
-import java.util.ArrayList;
 import java.util.List;
 //The Task class is in charge of creating task objects
 //which can be of different types: ToDos, Deadlines and Events
-//The Task class implements Serializable, and hence the tasklist generated by users
-//can be written into a text file Duke.txt when exiting the chatBot.
-//The tasklist can be opened again through reading the Duke.txt file the next time the chatBot is opened
 abstract class Task implements Serializable {
     protected String description;
     protected boolean isDone;
@@ -45,51 +41,5 @@ abstract class Task implements Serializable {
             System.out.print("\t" + i + "." + task.taskString());
             i++;
         }
-    }
-     static void saveTasks(String filePath, List<Task> tasks) {
-         ObjectOutputStream file = null;
-         try {
-             file = new ObjectOutputStream(new FileOutputStream(filePath));
-             file.writeObject(tasks);
-         } catch (IOException e) {
-             System.out.println(e.getMessage());
-         } finally {
-             if (file!= null) {
-                 try {
-                     file.close();
-                 } catch (IOException e) {
-                     throw new RuntimeException(e);
-                 }
-             }
-         }
-     }
-    static List<Task> loadTasks(String filePath) {
-        List<Task> tasks = new ArrayList<>();
-        ObjectInputStream file = null;
-        try {
-            @SuppressWarnings("unchecked")
-            //Unchecked warnings are yielded due to an unchecked type conversion
-            //readObject() returns an Object, but am casting it to List<Task>
-            //The compiler has no way to check if the object is of type List<Task>
-            //in compile time, resulting in an unchecked warning being yielded
-            File taskFile = new File(filePath);
-            if (taskFile.length() > 0) { //check if file is empty before attempting to read the file
-                file = new ObjectInputStream(new FileInputStream(taskFile));
-                tasks = (List<Task>) file.readObject();
-            }
-        } catch (ClassNotFoundException e) {
-            throw new RuntimeException(e);
-        } catch (IOException e) {
-            System.out.println(e.getMessage());
-        } finally {
-            if (file != null) {
-                try {
-                    file.close();
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
-            }
-        }
-        return tasks;
     }
 }
